Organize entities module
