+++ collect_app/src/main/java/org/odk/collect/android/entities/EntitiesRepositoryProvider.kt
@@ -2,7 +2,7 @@ package org.odk.collect.android.entities
 
 import org.odk.collect.android.projects.ProjectsDataService
 import org.odk.collect.android.storage.StoragePathProvider
-import org.odk.collect.entities.EntitiesRepository
+import org.odk.collect.entities.storage.EntitiesRepository
 import java.io.File
 
 class EntitiesRepositoryProvider(
