+++ collect_app/src/main/java/org/odk/collect/android/formentry/FormEntryUseCases.kt
@@ -15,8 +15,8 @@ import org.odk.collect.android.javarosawrapper.FormController
 import org.odk.collect.android.javarosawrapper.JavaRosaFormController
 import org.odk.collect.android.utilities.FileUtils
 import org.odk.collect.android.utilities.FormUtils
-import org.odk.collect.entities.EntitiesRepository
 import org.odk.collect.entities.LocalEntityUseCases
+import org.odk.collect.entities.storage.EntitiesRepository
 import org.odk.collect.forms.Form
 import org.odk.collect.forms.FormsRepository
 import org.odk.collect.forms.instances.Instance
