+++ collect_app/src/main/java/org/odk/collect/android/formentry/saving/FormSaver.java
@@ -5,7 +5,7 @@ import android.net.Uri;
 import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.tasks.SaveToDiskResult;
 import org.odk.collect.android.utilities.MediaUtils;
-import org.odk.collect.entities.EntitiesRepository;
+import org.odk.collect.entities.storage.EntitiesRepository;
 import org.odk.collect.forms.instances.InstancesRepository;
 
 import java.util.ArrayList;
