+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/ExperimentalPreferencesFragment.java
@@ -10,7 +10,7 @@ import androidx.fragment.app.FragmentActivity;
 import org.jetbrains.annotations.NotNull;
 import org.odk.collect.android.R;
 import org.odk.collect.androidshared.ui.ToastUtils;
-import org.odk.collect.entities.EntityBrowserActivity;
+import org.odk.collect.entities.browser.EntityBrowserActivity;
 
 public class ExperimentalPreferencesFragment extends BaseProjectPreferencesFragment {
