+++ collect_app/src/test/java/org/odk/collect/android/entities/InMemEntitiesRepositoryTest.kt
@@ -1,7 +1,7 @@
 package org.odk.collect.android.entities
 
-import org.odk.collect.entities.EntitiesRepository
-import org.odk.collect.entities.InMemEntitiesRepository
+import org.odk.collect.entities.storage.EntitiesRepository
+import org.odk.collect.entities.storage.InMemEntitiesRepository
 
 class InMemEntitiesRepositoryTest : EntitiesRepositoryTest() {
