+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/download/ServerFormDownloaderTest.java
@@ -25,8 +25,8 @@ import org.javarosa.xform.parse.XFormParser;
 import org.junit.Test;
 import org.odk.collect.android.formmanagement.FormMetadataParser;
 import org.odk.collect.android.formmanagement.ServerFormDetails;
-import org.odk.collect.entities.EntitiesRepository;
-import org.odk.collect.entities.InMemEntitiesRepository;
+import org.odk.collect.entities.storage.EntitiesRepository;
+import org.odk.collect.entities.storage.InMemEntitiesRepository;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.FormListItem;
 import org.odk.collect.forms.FormSource;
