+++ entities/src/main/java/org/odk/collect/entities/javarosa/EntityConstants.java
@@ -1,4 +1,4 @@
-package org.odk.collect.entities.javarosa.internal;
+package org.odk.collect.entities.javarosa;
 
 public class EntityConstants {
