+++ entities/src/main/java/org/odk/collect/entities/javarosa/EntityFormParser.java
@@ -1,20 +1,19 @@
-package org.odk.collect.entities.javarosa.internal;
+package org.odk.collect.entities.javarosa;
 
 import org.javarosa.core.model.data.IAnswerData;
 import org.javarosa.core.model.instance.FormInstance;
 import org.javarosa.core.model.instance.TreeElement;
-import org.odk.collect.entities.javarosa.EntityAction;
 import org.javarosa.xpath.expr.XPathFuncExpr;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ATTRIBUTE_BASE_VERSION;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ATTRIBUTE_CREATE;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ATTRIBUTE_DATASET;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ATTRIBUTE_ID;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ATTRIBUTE_UPDATE;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ELEMENT_ENTITY;
-import static org.odk.collect.entities.javarosa.internal.EntityConstants.ELEMENT_LABEL;
+import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_BASE_VERSION;
+import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_CREATE;
+import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_DATASET;
+import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_ID;
+import static org.odk.collect.entities.javarosa.EntityConstants.ATTRIBUTE_UPDATE;
+import static org.odk.collect.entities.javarosa.EntityConstants.ELEMENT_ENTITY;
+import static org.odk.collect.entities.javarosa.EntityConstants.ELEMENT_LABEL;
 
 public class EntityFormParser {
