+++ entities/src/main/java/org/odk/collect/entities/storage/EntitiesRepository.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.entities
+package org.odk.collect.entities.storage
 
 interface EntitiesRepository {
     fun save(vararg entities: Entity)
