+++ entities/src/test/java/org/odk/collect/entities/javarosa/EntityFormFinalizationProcessorTest.java
@@ -19,7 +19,6 @@ import org.javarosa.xform.util.XFormUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.odk.collect.entities.javarosa.internal.EntitiesExtra;
 
 public class EntityFormFinalizationProcessorTest {
