+++ entities/src/test/java/org/odk/collect/entities/javarosa/LocalEntitiesFileInstanceParserTest.kt
@@ -1,8 +1,11 @@
-package org.odk.collect.entities
+package org.odk.collect.entities.javarosa
 
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
+import org.odk.collect.entities.browser.EntityItemElement
+import org.odk.collect.entities.storage.Entity
+import org.odk.collect.entities.storage.InMemEntitiesRepository
 
 class LocalEntitiesFileInstanceParserTest {
